WIP general purpose line intersections for cast
We need to intersect view rays with non-grid aligned lines for doors
mid-animation, and all sprites. The lines are still aligned with x and y
axes, but not constrained to integer coordinates. The existing cast
algorithm relies on stepping in an integer grid, so add a general
purpose intersection function to be used instead.