Fix crash with enemy raycasting
memmove length to make space for the enemy hit in the array of station
hits was being calculated incorrectly, overflowing the hits array if the
start index was high enough (i.e. enough stations between player and
enemy). The t_cast_result struct containing the hits array is stored on
stack, so the overflow overwrote a stack protection cookie and caused a
crash with a stack smashing warning message.